==> Proyectos/EDA/Fase_2_transformacion/1_division_dataframes.ipynb <==
#pip install word2number

import pandas as pd
import numpy as np
from word2number import w2n


pd.set_option('display.max_columns', None)

df = pd.read_csv("../hr_raw_data_final.csv", index_col=0)
df.head(4)

# Función para convertir palabras a números
def convertir_a_numero(edad_palabra):
    # Si el valor ya es un número, lo dejamos tal cual
    if isinstance(edad_palabra, (int, float)):
        return edad_palabra
    try:
        # Si no es un número, tratamos de convertirlo a número
        return w2n.word_to_num(edad_palabra)
    except ValueError:
        # Si no se puede convertir, devolvemos None
        return None
# Aplicamos la función a la columna 'edad'
df['age'] = df['age'].apply(convertir_a_numero)

#Comprobamos que todas las edades estén en número y no en letra: 
df["age"].value_counts()

## Función para explorar el df 

def exploracion(df):

    df_info = pd.DataFrame()

    df_info["% nulos"] = round(df.isna().sum()/df.shape[0]*100, 2).astype(str)+"%"
    df_info["% no_nulos"] = round(df.notna().sum()/df.shape[0]*100, 2).astype(str)+"%"
    df_info["tipo_dato"] = df.dtypes
    df_info["num_valores_unicos"] = df.nunique()
    df_info

    print(f"""El DataFrame tiene {df.shape[0]} filas y {df.shape[1]} columnas.
Tiene {df.duplicated().sum()} datos duplicados, lo que supone un porcentaje de {round(df.duplicated().sum()/df.shape[0], 2)}% de los datos.

Hay {len(list(df_info[(df_info["% nulos"] != "0.0%")].index))} columnas con datos nulos, y son: 
{list(df_info[(df_info["% nulos"] != "0.0%")].index)}

y sin nulos hay {len(list(df_info[(df_info["% nulos"] == "0.0%")].index))} columnas y son: 
{list(df_info[(df_info["% nulos"] == "0.0%")].index)}


A continuación tienes un detalle sobre los datos nulos y los tipos y número de datos:""")
    
    display(df_info.head())

    print("Principales estadísticos de las columnas categóricas:")

    display(df.describe(include="O").T)

    print("Principales estadísticos de las columnas numéricas:")

    display(df.describe(exclude="O").T)

    return df_info

exploracion(df)

#Función para cambiar comas a puntos en decimales: 

def cambiar_comas (columnas_numericas):
    try:
        return float(columnas_numericas.replace(",", ".").replace("$", ""))
    
    except: np.nan

#Aplicamos la función en las columnas numéricas

columnas_numericas=['monthlyincome', "monthlyrate", "performancerating", "totalworkingyears", "worklifebalance", "yearsincurrentrole", "sameasmonthlyincome", "salary"]

for col in columnas_numericas:
    
    df[col] = df[col].apply(cambiar_comas)

#Comprobamos que las columnas numericas tengan el tipo correcto: 

df.dtypes

#Cambiamos la columna género a categorica: 
columnas_categoricas = ["gender"]

for col in columnas_categoricas:
    df[col] = df[col].astype("object")


#Comprobamos que el cambio de 'genre' se haya hecho: 

df.dtypes

# Dividimos en dos dataframes: 

df_numericos_1 = df.select_dtypes(include=['number'])  # Selecciona columnas numéricas
df_categoricos_2 = df.select_dtypes(include=['object'])  # Selecciona columnas categóricas

#Pasamos a CSV:
df_numericos_1.to_csv("numericos.csv", index=True)

df_categoricos_2.to_csv("categoricos.csv", index=True)
