==> Proyectos/EDA/Fase_2_transformacion/2_limpieza_categoricos.ipynb <==
import pandas as pd
import numpy as np


pd.set_option('display.max_columns', None)

- Tenemos que crear DataFrames de las columnas numéricas y categóricas por separado.
- Nos damos cuenta que algunas columnas que deberían ser numéricas están en categóricas y viceversa, como por ejemplo la coluna 'age' que está en categórica y pertenece a numérica
- Procedemos al cambio de las clumnas con sus respectivas tipologías

df_numericos = pd.read_csv('numericos.csv', index_col=0)

# Muestra los valores de 'age'
df_numericos['age'].value_counts()

df_categoricos=pd.read_csv("categoricos.csv", index_col=0)

#  Muestra las primeras 5 filas
df_categoricos.head()

- Vamos a empezar a sacar información de cada columna

# Muestra los valores
df_categoricos["roledepartament"].value_counts()

df_categoricos["jobrole"].value_counts()

# CAMNBIO DE NÚMEROS A LETRAS 
diccionario_gender={0: "M", 1: "F"}

# Se realzia con un mapeo
df_categoricos["gender"]=df_categoricos["gender"].map(diccionario_gender)

# Se comprueba
df_categoricos.head()

df_categoricos["remotework"].value_counts()

df_categoricos['remotework'] = df_categoricos['remotework'].replace({'0': 'no', '1': 'yes', 'True': 'yes', 'False': 'no', 'Yes': 'yes'})

df_categoricos['businesstravel'].value_counts()

#quitar "-" y "_" de la columna businesstravel: 
df_categoricos["businesstravel"] = df_categoricos["businesstravel"].str.replace('_', ' ', regex=False).str.replace('-', ' ', regex=False)

df_categoricos.head()

## Cambiamos a minuscula columna jobrole
df_categoricos['jobrole'] = df_categoricos['jobrole'].str.lower()

df_categoricos.head()


df_categoricos['maritalstatus'].value_counts()

## Cambiamos 'lo que esta mal escrito en columna de marital status'
df_categoricos['maritalstatus'] = df_categoricos['maritalstatus'].map({
    'Married': 'Married',
    'divorced': 'Divorced',
    'Single': 'Single',
    'Divorced': 'Divorced'
})

df_categoricos['maritalstatus'].value_counts()

df_categoricos['over18'] = df_categoricos['over18'].replace({'Y':'Yes', 'Yes':'Yes'})

df_categoricos['over18'].value_counts()

## Cambiamos a minuscula columna roledepartment 

df_categoricos['roledepartament'] = df_categoricos['roledepartament'].str.lower()

df_categoricos.head()

#Guardamos DF limpio en CSV
df_categoricos.to_csv("categoricos_limpio.csv", index=True)
