==> Proyectos/EDA/Fase_2_transformacion/3_limpieza_numericos.ipynb <==
import pandas as pd
import numpy as np


pd.set_option('display.max_columns', None)

df_numericos = pd.read_csv('numericos.csv', index_col=0)

df_categoricos=pd.read_csv("categoricos.csv", index_col=0)

# Mostrar las ultimas 5 filas
df_categoricos.tail()

# Mostrar las ultimas 5 filas
df_numericos.tail()

## Verificamos relacion entre columna 'salary' y 'monthlyincome'

df_numericos['is_equal'] = df_numericos.apply(lambda row: round(row['salary'] / 12, 2) == round(row['monthlyincome'],2)
                          if pd.notnull(row['salary']) and pd.notnull(row['monthlyincome'])
                          else None, axis=1)


df_numericos.tail()

# El número de valores nulos
df_numericos['is_equal'].isnull().sum()

# Número de valores nulos 
df_numericos['salary'].isnull().sum()

## Distance from home: cambiamos valores negativos a positivo: 

df_numericos['distancefromhome'] = df_numericos['distancefromhome'].abs()



df_numericos['distancefromhome'].value_counts()

df_numericos['employeenumber'].isnull().sum()

df_numericos['environmentsatisfaction'].value_counts()

df_numericos['performancerating'].value_counts()

df_numericos.columns

df_numericos['hourlyrate'].isnull().sum()

df_numericos['dailyrate'].isnull().sum()

## Tenemos que completar la columna de hourlyrate con los datos de daily rate (dailyrate / 8): 

df_numericos['hourlyrate'] = df_numericos['dailyrate'] / 8

#Comprobamos el cambio: 

df_numericos.head()

df_numericos['environmentsatisfaction'].unique()

## Cambiamos los valores de 'enviromentsatisfaction' (los que tienen mas de un digito deben tener uno solo)

labels = [1, 2, 3, 4] 

df_numericos['environmentsatisfaction'] = pd.cut(df_numericos['environmentsatisfaction'], bins=4, labels=labels, right=True) # Mostrar el DataFrame modificadoprint(df)

df_numericos['environmentsatisfaction'].value_counts()


df_numericos['is_equal'] = df_numericos.apply(lambda row: round(row['salary'] / 12, 2) == round(row['monthlyincome'],2)
                          if pd.notnull(row['salary']) and pd.notnull(row['monthlyincome'])
                          else None, axis=1)

df_numericos

df_numericos.head()

## Tenemos que completar los 'NaN' de monthlyincome con la división de salary / 12 
# Solo actualizar las filas de 'monthlyincome' donde sea NaN
df_numericos['monthlyincome'] = df_numericos.apply(
    lambda row: row['salary'] / 12 if pd.notnull(row['salary']) and pd.isnull(row['monthlyincome']) else row['monthlyincome'],
    axis=1
)


## Tenemos que completar los 'NaN' de salary con la multiplicacion de monthlyincome * 12 
# Solo actualizar las filas de 'salary' donde sea NaN
df_numericos['salary'] = df_numericos.apply(
    lambda row: row['monthlyincome'] * 12 if pd.notnull(row['monthlyincome']) and pd.isnull(row['salary']) else row['salary'],
    axis=1
)

df_numericos['monthlyincome'] = round(df_numericos['monthlyincome'],2) ##redondeamos monthly income a dos decimales. 

df_numericos.head()

df_numericos.tail()

## pasamos a CSV: 

df_numericos.to_csv("numericos_limpio_1.csv", index=True)
