==> Proyectos/EDA/Fase_2_transformacion/4_columnas_eliminadas.ipynb <==
import pandas as pd
import numpy as np


pd.set_option('display.max_columns', None)

df_numericos_limpio = pd.read_csv('numericos_limpio_1.csv', index_col=0)
df_numericos_limpio.head(1)


df_categoricos_limpio = pd.read_csv('categoricos_limpio.csv', index_col=0)
df_categoricos_limpio.head(10)

df_categoricos_limpio.info()

df_numericos_limpio.info()

df_unido = pd.concat([df_categoricos_limpio, df_numericos_limpio], axis=1)

df_unido.head(3)

df_unido.columns

## Eliminamos columnas innecesarias: 

columnas_eliminar = ['over18', 'maritalstatus', 'dailyrate', 'hourlyrate', 'sameasmonthlyincome', 'numberchildren', 'datebirth', 'roledepartament', 'standardhours','employeecount', 'yearsincurrentrole', 'totalworkingyears', 'is_equal']


df_unido = df_unido.drop(columns=columnas_eliminar)

df_unido.columns

df_unido.head()

df_unido.to_csv("unidos_limpio_1.csv", index=True)
