==> Proyectos/EDA/Fase_2_transformacion/5_data_junta.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', None)

df = pd.read_csv("unidos_limpio_1.csv", index_col=0)

df.head()

df.isna().sum()[df.isna().sum() > 0]

#Comprobar el porcentaje de nulos.
lista_columnas = ["businesstravel", "department", "educationfield", "overtime", "performancerating", "worklifebalance"  ]

for col in lista_columnas:
    procentajes_nulos = (df[col].isna().sum() / len(df)) * 100
    print(f'La columna {col.upper()} {df[col].isnull().sum()} valores nulos lo que es igual a {procentajes_nulos.round()} %')

df['educationfield'] = df['educationfield'].fillna('Unknown')

df["educationfield"].isna().sum()

valor_moda = df["overtime"].mode()[0]

df["overtime"].fillna(valor_moda)


df.head(10)

df["overtime"].isna().sum

df.head()

mediana = df["performancerating"].median()
df["performancerating"] = df["performancerating"].fillna(mediana)

df["performancerating"] = df["performancerating"].astype(int)

df["worklifebalance"].describe()[["mean", "50%"]]

mediana = df["worklifebalance"].median()
df["worklifebalance"] = df["worklifebalance"].fillna(mediana)

df["worklifebalance"] = df["worklifebalance"].astype(int)

df["worklifebalance"].value_counts()

df['businesstravel'] = df['businesstravel'].fillna('non travel')

df["businesstravel"].unique()

df['department'] = df['department'].fillna('Unknown')

df.head()

lista_minus = ["attrition","department","educationfield","overtime"]

for columna in lista_minus: 
    df[columna] = df[columna].str.lower()

df

df.columns

df_duplicados = df[df[['attrition', 'businesstravel', 'department', 'educationfield', 'gender',
       'jobrole', 'overtime', 'remotework', 'age', 'distancefromhome',
       'education', 'employeenumber', 'environmentsatisfaction',
       'jobinvolvement', 'joblevel', 'jobsatisfaction', 'monthlyincome',
       'monthlyrate', 'numcompaniesworked', 'percentsalaryhike',
       'performancerating', 'relationshipsatisfaction', 'stockoptionlevel',
       'trainingtimeslastyear', 'worklifebalance', 'yearsatcompany',
       'yearssincelastpromotion', 'yearswithcurrmanager', 'salary']].duplicated(keep=False)]

df_duplicados["employeenumber"].sort_values()

# Eliminar filas duplicadas en todo el DataFrame
df_duplicados = df.drop_duplicates()


df_duplicados.to_csv("data_limpia_1.csv", index=True)

df_duplicados.shape
